refactor(backend): introduced backend::trace module
